pyproject.toml
[project]
name = "karten-cli"
version = "0.0.2"
description = "A simple CLI application for creating vocabulary flashcards using generative AI, with functionality for bulk flashcard creation from Kindle dictionary lookups"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.8",
    "google-genai>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
karten="karten.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[build-system]
requires = ["setuptools"]

[tool.setuptools.packages]
find={}